Revert "created error check: cart quantity cannot exceed item stock"
This reverts commit 345bca4fdb81f5ec195b3784b9f636116bb2b926.

+++ backend/superprice/src/main/java/com/example/superprice/model/Product.java
@@ -1,4 +1,4 @@
 package com.example.superprice.model;
 
-public record Product(Long id, String name, String description, String store, String imageURL, Long price, int quantity) {
+public record Product(Long id, String name, String description, String store, String imageURL, Long price, Long quantity) {
 }
